requirements.txt
discord.py~=1.6.0
DateTime~=4.3
aiosqlite~=0.16.1
aiozaneapi~=1.5
validators~=0.18.2
PyYAML~=5.3.1
asyncio~=3.4.3
humanize~=3.1.0
psutil~=5.8.0
pur~=5.3.0
aiohttp~=3.7.3
Pillow~=8.1.0
aiofiles~=0.6.0
aiohttp~=3.7.3
wavelink~=0.9.6
asyncpraw~=7.1.0
pytz~=2020.1
wikipedia~=1.4.0
gTTS~=2.2.1
